Refactoring (created _utils_clustering)
